encryption: use generated nonces
Currently, we always re-use the same nonce based on the sequence count,
such that block 1 will use a nonce of "1", block 2 will use a nonce of
"2", etc. This effectively breaks the encryption scheme in use.

Instead of using a deterministic nonce based only on the block number,
we can also use a deterministic nonce based on the block count and
plain text. To do that, we use two separate keys K_enc and K_nonce,
derived from a master key via a key derival function, where K_enc is
used as the secret encryption key and K_nonce is used as the secret
key for the hash function. The nonce is then generated as follows,
assuming an block index i and data d:

 nonce_(i, d) = hash(i||d)

This nonce is then prepended to the generated block in plain and used to
encrypt respectively decrypt the block, together with key K_enc.

Even though this nonce is public (and needs to be, otherwise decyption
would be impossible), this key cannot be generated by anybody not
knowing the plain text. As it uses a hash, it furthermore does not
deduct any information about the data itself.