config.mk: move BSD/POSIX defines to CPPFLAGS
The CPPFLAGS are generally used for things like defines or includes
instead of using the CFLAGS variable. This is so that the CFLAGS
variable can be overridden, e.g. to include additional diagnostics or
the optimization level, while still maintaining the set of flags
necessary to build.

Fix the config.mk file to adhere to this rule by moving both the
"-D_BSD_SOURCE" and "-D_POSIX_C_SOURCE" defines to the CPPFLAGS
variable.