config.mk: avoid amending build flags
To build the project with a custom set of build flags, one can usually
either pass in the CFLAGS variable via an environment variable or by
specifying it after the make(1) command. The assumption is that this
will then completely override the set of build flags provided by the
project itself.

Instead of replacing the CFLAGS, we instead amend the passed in flags to
our usual set of flags. As previously explained, this is the wrong thing
to do. Fix this by instead using conditional assignment of our own
values, such that we will never clobber provided CFLAGS.