Makefile: use suffixes instead of special %.o syntax
The special target and dependency syntax "%.o: %.c" is not specified by
POSIX and thus does not work on systems which have a stricter make(1)
implementation, e.g. the BSDs. Replace it instead by using rules with
the special .SUFFIXES target, which are properly specified and thus
portable to POSIX conform systems.