config.mk: avoid use of `shell` for pkg-config
The `shell` builtin is a non-portable construct that is only understood
by GNU make. One example of systems not using GNU make by default are
the BSDs, which usually have their own flavor of make that more strictly
adheres to the POSIX specification. As they do not provide the `shell`
builtin, the pkg-config command is never executed and the builds are
thus broken.

Fix this by instead using "real" shell substitution using backticks. We
cannot use newer `$(cmd)` style substitution as make(1) uses this syntax
itself to perform variable substitution.