tests.sh: fix comparison of byte counts on OpenBSD
When testing whether output generated by the encryption algorithm always
is a multiple of the block size, we use wc(1) to test how many bytes the
output actually has and then compare it with the expected number of
bytes. This test fails on OpenBSD though, as `wc -c` will output a
leading space and thus the comparison will fail.

Fix the issue by using xargs(1) to strip spaces.