test.sh: fix non-portable use of head(1)
When veryfing that block lists without a trailer will cause git-cat to
fail, we use `head -n-1` to write everything but the last line
containing the trailer. The "-n-1" syntax is not specified by POSIX and
not implemented by OpenBSD's head(1).

As the test only generates two lines in total, simply change the
invocation to use the portable `head -n1` syntax to just output the
first line.