Makefile: avoid usage of non-standard patsubst function
We compute the set of manpages by simply appending ".1" to each of our
programs via the patsubst function. As this is a GNU extension only, it
fails the installation on OpenBSD by generating an emtpy set of
MANPAGES.

Fix the error by usingt the standard right-anchored substitution syntax
"MANPAGES=${PROGRAMS:=.1}, which does what we want it to do.