travis: provide key URL for Bintray
Due to a recently reported MITM RCE via Travis' APT package integration,
the build service has changed the default invocation of apt to not use
'--force-yes'. This causes our builds to fail, as we require packages
hosted via Bintray which are not signed by keys which are part of the
official Trusty distribution. As we do not force-install them anymore,
installing these packages will now fail.

Fix the issue by providing Bintray's key to the APT addon.