store: write blocks into temporary file first
Right now, we directly write blocks at the final path. While we try hard
to clean up after us in case where writing fails, such that we do not
leave a partly written block, this may not be possible in exceptional
circumstances (like e.g. SIGTERM or a complete crash of the host). This
is bad, as further writes to the block store will now assume that the
block with this specific hash has already been written.

Fix this by using the common pattern of writing into a temporary file,
followed by renaming that temporary file to the actual location.
Ideally, we'd be able to use mkstemp(3P). But we try hard to not care
about the actual location of the block store, and instead use file
descriptor relative function calls like openat(3P). Instead, we now just
use a ".tmp" suffix for the temporary files. While this will create
collisions as soon as a previous run of gob-chunk(1) has failed, it
still allows us to make better guarantees about consistency of each
blob.