common: always require hash string length
The length parameter of the `hash_from_str` function is currently
optional -- if it is not provided, we will derive it by using
strlen(3P). Currently, there are two callers that do not provide the
length, and in fact both already has the length available. So instead,
pass in the length of the string and let `hash_from_str` error out in
cases where no length was provided.