common: implement open-coded hex conversion
libsodium is currently used for both hash computations and also for
conversion from hex to binary representation and vice versa. As we want
to get rid of the libsodium dependency to make gob completely
self-hosted, remove the calls to `sodium_hex2bin` and `sodium_bin2hex`
as a first step to make that future happen.