common: implement error checking for stdout
When writing to stdout fails gob will not notice this right now, leading
to potential loss of e.g. the list of chunks stored by gob-chunk(1).
Introduce a new function `close_stdout` taken from gnulib that is being
executed at exit that checks the standard streams for errors and closes
them. If any error occurs or has occurred with stdout, then it will
print an error message and return an error code.