tests: add test case for invalid stdout
Verify that gob-chunk will error if stdout is redirected to an invalid
block device. The test makes use of /dev/full, which isn't accessible on
OSX. We thus add the capability to specify test prerequisites to a test.
If a prerequisite is not fulfilled, then the test is marked as skipped.