tests: drop useless quotes after assert_{success,failure}
The quotes around arguments to `assert_{success,failure}` were
supposedly needed to get redirection right in the spawned subcommands.
But in fact, the shell does get this right when using redirection for a
function and thus the quoting is not required.

Remove the quotes to make the tests easier to read.