common: fix warn() calling `exit`
The major difference between `die` and `warn` is that the former one
will call `exit`, while the latter one does. At least in theory, but in
fact `warn` calls `exit(1)` due to a copy-paste error.

Fix this by removing the call to `exit(1)` in `warn`. The fix causes our
current tests to fail because `assert_failure` has been checking for an
exit code of "1", while the change now causes gob-fsck(1) to exit "-1".
Fix this by checking for the error code not being "0" in
`assert_failure`.