common: fix return value of `store_read`
The function `store_read` returns the size of the block that has just
been read. To be consistent with all other functions that are similar to
that, we should return an `ssize_t` from this function to represent the
read block's size. This also fixes a warning about implicit conversion
losing integer precision.