gob-init: implement command to initialize stores
Currently, a blob store is automatically being created whenever
`store_open()` is called with a non-existing target path. Thus, if
calling e.g. git-chunk(1) with a non-existing directory, the command
would have just created the target and pretended everything to be fine.
This is a bit problematic and may hide errors like the wrong path being
passed to git-chunk(1). Even more problematic is that gob-fsck(1) would
create the target directory if it didn't exist.

Remove automatic creation of blob stores to fix this. Create a new
command gob-init(1) that sets up a repo and seeds its "version" file.