fsck: fix undefined behaviour when closing storefd
During the fsck, we call fdopendir(3P) to open the store's file
descriptor as a `DIR *` in order to iterate over it with readdir(3P).
Afterwards, we use both closedir(3P) on the `DIR *` and close(3P) on the
store's file descriptor. Quoting from fdopendir(3P):

    Upon successful return from fdopendir(), the file descriptor is
    under the control of the system, and if any attempt is made to close
    the file descriptor, or to modify the state of the associated
    description, other than by means of closedir(), readdir(),
    readdir_r(), rewinddir(), or seekdir(), the behavior is undefined.
    Upon calling closedir() the file descriptor shall be closed.

Summarized, closing both the dir and the file descriptor is undefined
behaviour.

Fix the issue by calling dup(3P) on the store's file descriptor first.